Refactored Migration scripts as well as some other things according to PR comments
