+++ retrograde-app-shared/src/main/java/com/swordfish/lemuroid/lib/library/db/dao/Migrations.kt
@@ -49,51 +49,19 @@ object Migrations {
     val VERSION_9_10: Migration =
         object : Migration(9, 10) {
             override fun migrate(database: SupportSQLiteDatabase) {
-                database.execSQL(
-                    "ALTER TABLE games ADD COLUMN customName TEXT DEFAULT NULL",
-                )
-                database.execSQL(
-                    "ALTER TABLE games ADD COLUMN customCoverUri TEXT DEFAULT NULL",
-                )
-
-                database.execSQL("DROP TRIGGER IF EXISTS games_bu")
-                database.execSQL("DROP TRIGGER IF EXISTS games_bd")
-                database.execSQL("DROP TRIGGER IF EXISTS games_au")
-                database.execSQL("DROP TRIGGER IF EXISTS games_ai")
-
-                database.execSQL(
-                    """
-                    CREATE TRIGGER games_bu BEFORE UPDATE ON games BEGIN
-                      DELETE FROM fts_games WHERE docid=old.id;
-                    END;
-                    """.trimIndent(),
-                )
-                database.execSQL(
-                    """
-                    CREATE TRIGGER games_bd BEFORE DELETE ON games BEGIN
-                      DELETE FROM fts_games WHERE docid=old.id;
-                    END;
-                    """.trimIndent(),
-                )
-                database.execSQL(
-                    """
-                    CREATE TRIGGER games_au AFTER UPDATE ON games BEGIN
-                      INSERT INTO fts_games(docid, title) VALUES(new.id, COALESCE(new.customName, new.title));
-                    END;
-                    """.trimIndent(),
-                )
-                database.execSQL(
-                    """
-                    CREATE TRIGGER games_ai AFTER INSERT ON games BEGIN
-                      INSERT INTO fts_games(docid, title) VALUES(new.id, COALESCE(new.customName, new.title));
-                    END;
-                    """.trimIndent(),
-                )
-
-                database.execSQL("DELETE FROM fts_games")
-                database.execSQL(
-                    "INSERT INTO fts_games(docid, title) SELECT id, COALESCE(customName, title) FROM games",
-                )
+                migrateGamesTableToVersion10(database)
+                GameSearchDao.migrateFtsForVersion10(database)
             }
         }
+
+    private fun migrateGamesTableToVersion10(
+        database: SupportSQLiteDatabase,
+    ) {
+        database.execSQL(
+            "ALTER TABLE games ADD COLUMN customName TEXT DEFAULT NULL",
+        )
+        database.execSQL(
+            "ALTER TABLE games ADD COLUMN customCoverUri TEXT DEFAULT NULL",
+        )
+    }
 }
